refactor: change ui/FileBrowser folder to ui/BrowsePage
- and update related imports
- more clearly describes the purpose of the ui components contained in the folder
