+++ src/components/ui/BrowsePage/ZarrMetadataTable.tsx
